add style preludes
also adds the IsEndo class, which is not actually necessary for style preludes. however, the need for IsEndo has come up in the past as well; even though it always turns out to be a false need i figure it's not too much work to keep around.

there's also the following benefit of clarity. currently to compose endos we write `<>`. but both functionlike composition and pointwise-append form monoids, so 'a priori' it's unclear which of these `<>` is. with IsEndo we give these different names; namely, `.<<` and `.<>`