introduce MationT for internal usage
before this commit it was possible to write a "onClick" prop, which is triggered when the target element is clicked, but not possible to write a "onClickElsewhere" prop, which is triggered when any element *but* the target is clicked

the reason for this asymmetry was the following. since onClick corresponds to a built-in HTML event listener, but onClickElsewhere does not, they were represented differently within the framework. the first was represented as a so-called "listener" and given type `DomNode -> Mation m s`, and the second was represented as a "fixup" (≈ "custom prop") and with type `DomNode -> Effect { restore :: Effect Unit }`.

when it came time for the framework to render an application containing listeners/fixups, they were treated differently. to make use of a listener (of type `DomNode -> Mation m s`), the framework supplied to the listener the application `Step s` stepper in order to transform the listener into a `DomNode -> Effect Unit` for consumption by the patching code. on the contrary, fixups (of type `DomNode -> Effect { restore :: Effect Unit }`) were passed to the patching code as-is *without ever recieving* the application `Step s`.

in short, the issue was that listeners got access to the application `Step s` stepper function whereas fixups (≈ "custom props") did not. as such, onClick could update application state whereas onClickElsewhere could not.

to rectify this asymmetry, this commit does the following. it first generalizes the type `Mation m s ~ Step s -> m Unit` to the monadic type `MationT m s a ~ Step s -> m a`. then it changes the listener representation from `DomNode -> Mation m s` to `DomNode -> MationT m s Unit`, which is equivalent. it additionally changes the fixup representation from `DomNode -> Effect { restore :: Effect Unit }` to (roughly an equivalence of) `DomNode -> MationT m s { restore :: MationT m s Unit }`. this change means that fixups now have access to the application `Step s`, just as listeners do. note that the generalization of `Mation` to `MationT` was necessary because fixups have return values (unlike listeners)

it should be noted that fixups having access to the application `Step s` is very powerful. it means an API user could potentially kick off a long-running computation (such as an API call) at *render* time, and have it update the application state (ie, not be useless). warnings have been placed on the relevant fixup constructors: a fixup should make use of a `Step s` with care and caution

this commit contains some collateral changes to the codebase besides what's been discussed. following is an approximate overview of changes made to the codebase with this commit

- `MationT` introduced

- `Mation` changed from a `newtype` to a type alias over `MationT`. several helper functions regarding `Mation` removed/modified

- fixups now live in `m` instead of `Effect` and are now given access to the application `Step s`

- introduction of a `Revertible` type which is relevant only for highly-technical details; namely, it allows us to avoid constraints on many API exports