capabilities wip: couple ref types to Effect before this change, ref types were indexed by monad, so you could have both
  ReadWrite Effect Int

and

  ReadWrite Aff Int

in practice, though, the ref monad type is always going to be Effect, since it's created by the Mation runtime (see the code comment).
so, for simplicity, we just couple the ref types to effect.

this change also removes the footgun of users writing code like

  Html m (SomeRefType m s)

which seems fine but should actually be

  Html m (SomeRefType Effect s)

and will cause trouble down the road.